requirements.txt
Flask
reportlab
python-docx
deep-translator